+++ src/intelli/gui/guidata.cpp
@@ -45,7 +45,7 @@ LocalStateContainer::LocalStateContainer(GtObject* parent) :
 {
     setObjectName(tr("local_states"));
 
-    // presence of struct inidcates node is collapsed
+    // presence of struct indicates that the node is collapsed
     GtPropertyStructDefinition structType{S_TYPE_ID};
     m_collapsed.registerAllowedType(structType);
 
@@ -53,29 +53,19 @@ LocalStateContainer::LocalStateContainer(GtObject* parent) :
 
     connect(&m_collapsed, &GtPropertyStructContainer::entryAdded,
             this, [this](int idx){
-                GtPropertyStructInstance& entry = m_collapsed.at(idx);
-                m_collapsedMap.insert(idx, entry.ident());
-                emit nodeCollapsedChanged(entry.ident(), true);
-                assert(m_collapsed.size() == (size_t)m_collapsedMap.size());
-            }, Qt::DirectConnection);
+        GtPropertyStructInstance& entry = m_collapsed.at(idx);
+        m_collapsedData.insert(idx, entry.ident());
+        emit nodeCollapsedChanged(entry.ident(), true);
+        assert(m_collapsed.size() == (size_t)m_collapsedData.size());
+    }, Qt::DirectConnection);
 
     connect(&m_collapsed, &GtPropertyStructContainer::entryRemoved,
             this, [this](int idx){
-                NodeUuid nodeUuid = m_collapsedMap.at(idx);
-                m_collapsedMap.removeAt(idx);
-                emit nodeCollapsedChanged(nodeUuid, false);
-                assert(m_collapsed.size() == (size_t)m_collapsedMap.size());
-        }, Qt::DirectConnection);
-}
-
-void
-LocalStateContainer::init()
-{
-    assert(m_collapsed.size() == (size_t)m_collapsedMap.size());
-    for (GtPropertyStructInstance const& e : m_collapsed)
-    {
-        emit nodeCollapsedChanged(e.ident(), true);
-    }
+        NodeUuid nodeUuid = m_collapsedData.at(idx);
+        m_collapsedData.removeAt(idx);
+        emit nodeCollapsedChanged(nodeUuid, false);
+        assert(m_collapsed.size() == (size_t)m_collapsedData.size());
+    }, Qt::DirectConnection);
 }
 
 void
@@ -85,10 +75,8 @@ LocalStateContainer::setNodeCollapsed(NodeUuid const& nodeUuid, bool collapsed)
 
     if (collapsed)
     {
-        m_collapsed.newEntry(S_TYPE_ID,
-                             std::next(m_collapsed.begin(), m_collapsed.size()),
-                             nodeUuid);
-        assert(m_collapsed.size() == (size_t)m_collapsedMap.size());
+        m_collapsed.newEntry(S_TYPE_ID, nodeUuid);
+        assert(m_collapsed.size() == (size_t)m_collapsedData.size());
         return;
     }
 
@@ -99,16 +87,16 @@ LocalStateContainer::setNodeCollapsed(NodeUuid const& nodeUuid, bool collapsed)
     if (iter == m_collapsed.end()) return;
 
     m_collapsed.removeEntry(iter);
-    assert(m_collapsed.size() == (size_t)m_collapsedMap.size());
+    assert(m_collapsed.size() == (size_t)m_collapsedData.size());
 }
 
 bool
 LocalStateContainer::isNodeCollapsed(NodeUuid const& nodeUuid) const
 {
-    assert(m_collapsed.size() == (size_t)m_collapsedMap.size());
+    assert(m_collapsed.size() == (size_t)m_collapsedData.size());
     return std::find_if(m_collapsed.begin(), m_collapsed.end(),
                         [&nodeUuid](GtPropertyStructInstance const& e){
-                            return e.ident()  == nodeUuid;
-                        }) != m_collapsed.end();
+        return e.ident()  == nodeUuid;
+    }) != m_collapsed.end();
 }